feat: Add support for associations and UML class types
- Implement one-way associations
- Implement generalizations
- Implement compositions
- Support all types of classes in DrawIO UML, including enumerations (Class(1..55)
-refactor name of function and folder
