+++ besser/BUML/notations/structuralDrawIO/__init__.py
@@ -1,0 +1,1 @@
+from .structural_drawio_to_buml import *
