--- besser/BUML/notations/structuralDrawioXML/__init__.py
@@ -1,1 +1,0 @@
-from .drawio_to_buml import *
