--- besser/BUML/notations/structuralPlantUML/__init__.py
@@ -1,6 +1,0 @@
-from .plantuml_to_buml import *
-from .plantUML_buml_listener import *
-from .PlantUMLParser import *
-from .PlantUMLLexer import *
-from .PlantUMLListener import *
-from .plantUML_buml_listener import *
